% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_fun.R
\name{autotune_cissvae}
\alias{autotune_cissvae}
\title{Autotune CISS-VAE hyperparameters with Optuna
...}
\usage{
autotune_cissvae(
  data,
  index_col = NULL,
  clusters,
  save_model_path = NULL,
  save_search_space_path = NULL,
  n_trials = 20,
  study_name = "vae_autotune",
  device_preference = "cuda",
  show_progress = FALSE,
  optuna_dashboard_db = NULL,
  load_if_exists = TRUE,
  seed = 42,
  verbose = FALSE,
  num_hidden_layers = c(1, 4),
  hidden_dims = c(64, 512),
  latent_dim = c(10, 100),
  latent_shared = c(TRUE, FALSE),
  output_shared = c(TRUE, FALSE),
  lr = c(1e-04, 0.001),
  decay_factor = c(0.9, 0.999),
  beta = 0.01,
  num_epochs = 10,
  batch_size = 64,
  num_shared_encode = c(0, 1, 3),
  num_shared_decode = c(0, 1, 3),
  refit_patience = 2,
  refit_loops = 100,
  epochs_per_loop = 10,
  reset_lr_refit = c(TRUE, FALSE)
)
}
\arguments{
\item{num_hidden_layers}{Numeric(2) vector: (min, max) for # hidden layers.}

\item{num_shared_encode}{Numeric vector: categorical # shared encoder layers.}

\item{num_shared_decode}{Numeric vector: categorical # shared decoder layers.
...}
}
\description{
Autotune CISS-VAE hyperparameters with Optuna
...
}
