% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_venv.R
\name{create_cissvae_env}
\alias{create_cissvae_env}
\title{Create or reuse a CISSVAE Python virtual environment}
\usage{
create_cissvae_env(
  envname = "cissvae_environment",
  install_python = FALSE,
  python_version = "3.10"
)
}
\arguments{
\item{envname}{Name of the virtual environment to create/use.}

\item{install_python}{Logical; if TRUE, install Python if none found.}

\item{python_version}{Python version string (major.minor), if you need to install.}
}
\description{
Create or reuse a CISSVAE Python virtual environment
}
