% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{dni}
\alias{dni}
\title{Example dni matrix for demo of do_not_impute_matrix}
\format{
A dataframe:
\describe{
\item{do_not_impute_matrix}{A mock do_not_impute_matrix dataframe}
}
}
\source{
Imported from \code{data_raw/dni.csv}
}
\usage{
dni
}
\description{
A sample do_not_impute_matrix (dataframe)
}
\examples{
data(dni)
}
\keyword{datasets}
