% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ign_plt_silh.R
\name{plot_silhouette_beeswarm}
\alias{plot_silhouette_beeswarm}
\title{Plot a beeswarm of silhouette scores}
\usage{
plot_silhouette_beeswarm(silhouettes, sample_ids = NULL)
}
\arguments{
\item{silhouettes}{Numeric vector of silhouette scores.}

\item{sample_ids}{Optional character vector of sample identifiers. If omitted,
samples will be numbered sequentially.}
}
\value{
A \code{ggplot} object showing a beeswarm of silhouette scores.
}
\description{
Plot a beeswarm of silhouette scores
}
\examples{
\dontrun{
  sil <- run_cissvae(df, return_silhouettes = TRUE)$silhouettes
  plot_silhouette_beeswarm(sil)
}
}
